refactor(monitoring): split summary contracts and archive history
