+++ src/foehncast/monitoring/__init__.py
@@ -18,10 +18,14 @@ from foehncast.monitoring.pipeline_metrics import (
     emit_feature_pipeline_run_summary,
     emit_training_pipeline_run_summary,
     feature_pipeline_stage_overview,
+    feature_pipeline_summary_history_paths,
     feature_pipeline_summary_path,
+    read_feature_pipeline_run_summary_history,
     read_feature_pipeline_run_summary,
+    read_training_pipeline_run_summary_history,
     read_training_pipeline_run_summary,
     training_pipeline_stage_overview,
+    training_pipeline_summary_history_paths,
     training_pipeline_summary_path,
 )
 from foehncast.monitoring.prediction_log import (
@@ -53,6 +57,7 @@ __all__ = [
     "emit_feature_pipeline_run_summary",
     "emit_training_pipeline_run_summary",
     "feature_pipeline_stage_overview",
+    "feature_pipeline_summary_history_paths",
     "feature_pipeline_summary_path",
     "prediction_event_log_path",
     "push_drift_metrics",
@@ -60,10 +65,13 @@ __all__ = [
     "read_prediction_event_log",
     "read_prediction_history",
     "read_prediction_log",
+    "read_feature_pipeline_run_summary_history",
     "read_feature_pipeline_run_summary",
+    "read_training_pipeline_run_summary_history",
     "read_training_pipeline_run_summary",
     "render_online_compose_sync_prometheus_metrics",
     "render_prediction_log_prometheus_metrics",
     "training_pipeline_stage_overview",
+    "training_pipeline_summary_history_paths",
     "training_pipeline_summary_path",
 ]
